Add Neural Harmonic Textures (NHT) feature mode for 3DGRT and 3DGUT
Introduces NHT as a learned alternative to spherical harmonics for radiance
representation. Key changes:

Feature rendering (CUDA/C++):
- Forward feature buffer uses fp16 when feature_output_half=true (halves
  memory bandwidth), fp32 otherwise; controlled by FEATURE_OUTPUT_HALF flag
- Gradient buffer is always fp32: radianceDensityGradientPtr changed from
  TRadianceDensityElem* to float* in renderBackward, the CUDA backward kernel,
  and splatRaster.cpp — matching the reference NHT implementation (gsplat
  _wrapper.py line 2899: v_render_colors = v_fused_renders.float())

Python training:
- _Fp16ToFp32SafeGrad: custom autograd function for fp16 feature maps whose
  backward returns fp32 (avoids lossy fp32→fp16→fp32 gradient roundtrip)
- tracer.py: .float() on ray_radiance_density_grd before trace_bwd to
  guarantee fp32 gradient regardless of PyTorch autograd dtype inference
- apply_feature_decoder: uses _Fp16ToFp32SafeGrad when feature map is fp16
- mcmc.py: guard against torch.multinomial crash when alive_idxs is empty

Config alignment with reference (neural-harmonic-textures/gsplat):
- All NHT app configs: enable use_opacity and use_scale losses (lambda values
  were already set but flags were not enabled)

+++ threedgut_tracer/tracer.py
@@ -258,7 +258,7 @@ class Tracer:
                 sensor_poses.T_world_sensors[0],
                 sensor_poses.T_world_sensors[1],
                 ray_radiance_density,
-                ray_radiance_density_grd,
+                ray_radiance_density_grd.float(),  # gradient buffer is always fp32 (see rayPayloadBackward.cuh)
                 ray_hit_distance,
                 ray_hit_distance_grd,
             )
